src/dev_requirements.txt
-r requirements.txt

bandit==1.7.0
flake8==3.9.2
flake8-blind-except==0.2.0
flake8-broken-line==0.3.0
flake8-bugbear==21.9.2
flake8-builtins==1.5.3
flake8-comprehensions==3.6.1
flake8-django==1.1.2
flake8-eradicate==1.1.0
flake8-executable==2.1.1
mypy==0.910
pytest==6.2.5
pytest-cov==2.12.1
pytest-django==4.1.0
pytest-flake8==1.0.7
pytest-forked==1.3.0
pytest-mypy==0.8.1
pytest-xdist==2.4.0
responses==0.14.0
safety==1.10.3
types-pytz==2021.1.0
types-requests==2.25.1
black==21.9b0
invoke==1.6.0
